requirements/base.txt
# Core Django dependencies
Django>=5.2,<6.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.5.1
django-mptt>=0.14.0
django-cors-headers>=4.2.0
python-gettext>=4.1

# Build tools and packaging
setuptools>=65.0.0,<81.0.0

# Media handling
Pillow>=10.0.0

# HTTP Client
requests>=2.31.0

# JWT Support
PyJWT>=2.8.0

# Redis for caching and rate limiting
redis>=5.0.0
django-redis>=5.4.0

# API Documentation
drf-spectacular>=0.27.0

# Background tasks
celery>=5.3.0
redis>=5.0.0